chore(permission middleware): add permission middleware
	- add middleware to test the id provided by the user
	- add also test if user should have access to data
