+++ server/routes/router.js
@@ -2,17 +2,18 @@ import express from 'express';
 import userController from '../controllers/userController';
 import entryController from '../controllers/entryController';
 import auth from '../middlewares/auth';
+import permission from '../middlewares/permission';
 
 const router = express.Router();
 
 router.get('/api/v1/', userController.entryMessage);
 router.get('/api/v1/entries', auth, entryController.allEntries);
-router.get('/api/v1/entries/:entryId', auth, entryController.getEntry);
+router.get('/api/v1/entries/:entryId', auth, permission, entryController.getEntry);
+router.patch('/api/v1/entries/:entryId', auth, permission, entryController.modifyEntry);
+router.delete('/api/v1/entries/:entryId', auth, permission, entryController.deleteEntry);
 router.post('/api/v1/auth/signup', userController.createAccount);
 router.post('/api/v1/auth/signin', userController.Login);
 router.post('/api/v1/entries', auth, entryController.addEntry);
-router.patch('/api/v1/entries/:entryId', auth, entryController.modifyEntry);
-router.delete('/api/v1/entries/:entryId', auth, entryController.deleteEntry);
 
 
 export default router;
